Prompt if no client found (#14)
Instead of just saying the program couldn't find the client, it now prompts the user to select an existing java process to run on. This allows users using modded clients to use the autoclicker.

I also added RLCraft to the default clients it searches for and made it so it can expand to other clients easily in the future.